+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-08-04 20:15:41
-// Commit: docs(forms): update readme and API contracts for forms
+// Updated at: 2026-08-06 12:36:53
+// Commit: test(forms): implement widget integration tests for forms
 
 class formsComponentState {
-  final String commitRef = '3369';
-  final String lastMessage = 'docs(forms): update readme and API contracts for forms';
-  final DateTime timestamp = DateTime.parse('2026-08-04T20:15:41+05:00');
+  final String commitRef = '3391';
+  final String lastMessage = 'test(forms): implement widget integration tests for forms';
+  final DateTime timestamp = DateTime.parse('2026-08-06T12:36:53+05:00');
 }